[#3852] additional validation S5 targets; added type validation
